+++ include/aws/http/request_response.h
@@ -165,10 +165,8 @@ struct aws_http_request_options {
     bool manual_window_management;
 };
 
-typedef int(aws_transform_http_request_fn)(
-    struct aws_http_message *request,
-    struct aws_allocator *allocator,
-    void *context);
+typedef int(
+    aws_http_request_transform_fn)(struct aws_http_message *request, struct aws_allocator *allocator, void *context);
 
 /**
  * Initializes aws_http_request_options with default values.
